requirements.txt
flask==2.3.3
python-docx==0.8.11
ollama==0.4.7
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2 